optimization of drawing of particles
after profiling I found that draw circles is *really* slow (makes sense,
it draws a lot of triangles).
Instead of that, I am now rendering a single texture (2 triangles), and
scaling it as needed.
Now this allows for a lot more particles (more than make sense in such a
small canvas). I'll probably have to make the "world space" bigger than
the literal window size so that I can see the actual performance.